requirements.txt
absl-py==2.3.1
accelerate==1.9.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audiocraft==1.2.0
audioread==3.0.1
av==11.0.0
babel==2.17.0
bangla==0.0.5
beartype==0.21.0
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.39.4
botocore==1.39.4
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
cn2an==0.5.23
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
datasets==2.16.1
dateparser==1.1.8
decorator==5.2.1
demucs==4.0.1
Deprecated==1.2.18
dill==0.3.7
distlib==0.4.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
dora_search==0.1.12
easydict==1.13
einops==0.8.1
einx==0.3.0
elevenlabs==2.8.0
ema-pytorch==0.7.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
encodec==0.1.1
espeakng-loader==0.2.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
flashy==0.0.2
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2023.10.0
ftfy==6.3.1
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.45
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.59.0
grpcio-tools==1.59.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
interegular==0.3.3
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
kokoro==0.9.4
lameenc==1.8.1
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.1
lightning-utilities==0.14.3
llvmlite==0.43.0
lm-format-enforcer==0.10.11
loguru==0.7.3
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
misaki==0.9.4
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.15
muq==0.1.0
murmurhash==1.0.13
mutagen==1.47.0
networkx==2.8.8
nltk==3.9.1
nnAudio==0.3.3
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
omegaconf==2.3.0
onnxruntime==1.22.1
openunmix==1.3.0
orjson==3.11.0
packaging==25.0
pandas==1.5.3
parso==0.8.4
phonemizer==3.3.0
phonemizer-fork==3.3.2
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
proces==0.1.7
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
protos==0.1.0
psutil==7.0.0
pure_eval==0.2.3
py3langid==0.3.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.19.2
pykakasi==2.3.0
pymongo==4.13.2
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pyreadline3==3.5.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
replicate==1.0.7
requests==2.32.4
resemble-perth==1.0.1
retrying==1.4.1
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.26.0
ruff==0.12.4
s3transfer==0.13.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.11.4
segments==2.3.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
submitit==1.5.3
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.15.2
tomlkit==0.13.3
torch==2.1.0+cu121
torchaudio==2.1.0
torchdata==0.7.0
torchdiffeq==0.2.5
torchmetrics==1.7.4
torchtext==0.16.0
torchvision==0.16.0
tqdm==4.66.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.37.2
treetable==0.2.5
TTS==0.22.0
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
wandb==0.21.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.0.2
win32_setctime==1.2.0
wrapt==1.17.2
x-clip==0.14.4
x-transformers==2.5.2
xformers==0.0.22.post7
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
